fix(types): resolve mypy errors across pipeline
- AlignedToken.start_ms/end_ms: int -> float (NaN placeholders are
  intentional; _interpolate_nans fills them later)
- MediaInfo.path: str -> Path; update CLI construction site and JSON
  serializer to handle PurePath
- AlignmentBackendFactory.create: language widened to str | None to
  match config schema; existing ValueError path preserves spec
  contract that AlignmentStage converts to FAILED status
- TranscriptionStage: assert step.backend is not None after
  ConfigLoader._resolve_transcribing_backends fills defaults
- ConfigLoader: drop invalid 'msg' key from InitErrorDetails
- mypy: ignore_missing_imports for ffmpeg/ten_vad/silero_vad/
  torchaudio/transformers; add types-PyYAML dev dep
- post_processing.word_boundary: cast AlignedToken float ms back to
  int when building Subtitle (still int)

Also fix mlx_whisper backend test isolation: patch
importlib.import_module on the backend's namespace instead of
builtins.__import__ (the backend uses importlib.import_module which
bypasses the builtin hook).

+++ src/talking_parrot/models/transcription.py
@@ -14,8 +14,8 @@ class TranscriptionMetrics:
 @dataclass(frozen=True)
 class AlignedToken:
     word: str
-    start_ms: int
-    end_ms: int
+    start_ms: float
+    end_ms: float
     score: float
